+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - added keyboard shortcut `mod + e` for enabling / disabling the editor
   - added keyboard shortcuts `mod + arrows` for selecting widgets
   - added keyboard shortcut `mod + a` for selecting a widget's siblings and itself
+  - added keyboard shortcut `mod + shift + a` for cancelling selection
   - added multi widget selection with mouse
 
 ## 0.37.4